+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Extracted common code to [Cucumber Query](https://github.com/cucumber/query/tree/main) ([#30](https://github.com/cucumber/cucumber-junit-xml-formatter/pull/31), M.P. Korstanje)
 
 ## [0.3.0] - 2024-03-23
 ### Added